Use UIViewController for navigation events
